load js from HTTPS source
If the page is hosted with HTTPS enabled site, such as GitHub Pages, then the library will not be loaded because it is from insecure source. In order to solve this problem, external library needs to be loaded with HTTPS.

By the way, I think it is a best particle to use HTTPS connection if possible.